+++ packages/ripple-address-codec/.eslintrc.js
@@ -39,6 +39,7 @@ module.exports = {
     'jsdoc/require-throws': 'off',
     'jsdoc/require-jsdoc': 'off',
     'jsdoc/require-param': 'off',
+    'jsdoc/check-examples': 'off', // Not implemented in eslint 8
     'tsdoc/syntax': 'off',
     '@typescript-eslint/no-require-imports': 'off',
   },