+++ packages/xrpl/src/models/methods/accountObjects.ts
@@ -1,4 +1,4 @@
-import { AccountObjectType, LedgerIndex } from '../common'
+import { LedgerIndex } from '../common'
 import {
   Check,
   DepositPreauth,
@@ -11,6 +11,17 @@ import {
 } from '../ledger'
 
 import { BaseRequest, BaseResponse } from './baseMethod'
+
+type AccountObjectType =
+  | 'check'
+  | 'deposit_preauth'
+  | 'escrow'
+  | 'nft_offer'
+  | 'offer'
+  | 'payment_channel'
+  | 'signer_list'
+  | 'state'
+  | 'ticket'
 
 /**
  * The account_objects command returns the raw ledger format for all objects
@@ -65,8 +76,8 @@ type AccountObject =
   | Offer
   | PayChannel
   | SignerList
-  | Ticket
   | RippleState
+  | Ticket
 
 /**
  * Response expected from an {@link AccountObjectsRequest}.