+++ packages/xrpl/HISTORY.md
@@ -11,6 +11,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 * Fixed empty value condition for NFTokenMinter field in AccountSet transaction
 
 ### Added
+- `getNFTokenID` lets you get the NFTokenID after minting an NFT
 
 
 ### Changed
@@ -27,6 +28,7 @@ Wallet.fromMmnemonic()
 
 ### Added
 * Optional custom amount field to `fundWallet`.
+* Support for `disallowIncoming` account set flags (e.g. `asfDisallowIncomingTrustline`)
 
 ### Changed
 * Add support for Transaction objects in `verifyTransaction`